Plugins (dancingparticles): Rename t1font.cpp and t1font.h to font.cpp and font.h respectively.
