+++ libvisual-plugins/plugins/actor/dancingparticles/parameters.cpp
@@ -1,7 +1,7 @@
 #include <string.h>
 #include <libvisual/libvisual.h>
 #include "etoile.h"
-#include "t1font.h"
+#include "font.h"
 
 
 // or why use lex when..
